moves postfix main config above aliases
Is needed if you want to disable IPv6: The command newaliases will fail
if IPv6 is stilled enabled in main.cf but your server doesn't support
IPv6.

With this change the main.cf will be changed before the aliases are
created. I can't tell if there will be complications in other
situations/configurations.